requirements.txt
crewai
crewai-tools
langchain-community
langchain-google-genai
python-dotenv
pillow
streamlit